fix(webhooks): expose type, fix docs, verifyWebhookSignature helper
