+++ docs-v2/integrate/guides/receive-webhooks-from-an-api.mdx
@@ -27,7 +27,7 @@ When Nango receives a webhook from an external API, it notifies your app with an
 
 # Forward external webhooks to your app
 
-Nango automatically forwards all webhooks from external APIs to your app. 
+Nango automatically forwards all webhooks from external APIs to your app.
 
 Nango webhook forwards are POST requests with the following JSON body:
 ```json
@@ -42,7 +42,7 @@ Nango webhook forwards are POST requests with the following JSON body:
 
 # Handle external webhooks in syncs
 
-Additionally, [syncs](/understand/concepts/syncs) can be designed to process webhooks from external APIs. 
+Additionally, [syncs](/understand/concepts/syncs) can be designed to process webhooks from external APIs.
 
 Webhook-enabled syncs process relevant webhooks, resulting in the creation, update and deleting of the corresponding [records](/understand/concepts/syncs) in real-time.
 
@@ -53,7 +53,7 @@ Each modified record (following an external webhook) will trigger a separate web
     "providerConfigKey": "<INTEGRATION-ID>",
     "syncName": "<SYNC-NAME>",
     "model": "<MODEL-NAME>",
-    "responseResults": { "<MODEL-NAME>": { "added": 1, "updated": 0, "deleted": 0 } },
+    "responseResults": { "added": 1, "updated": 0, "deleted": 0 },
     "syncType": "WEBHOOK",
     "modifiedAfter": "<TIMESTAMP>"
 }
@@ -61,7 +61,7 @@ Each modified record (following an external webhook) will trigger a separate web
 
 Simply activate a webhook-enabled sync to benefit from these capabilities.
 
-You can check if a sync supports external webhooks by selecting your integration, going to the _Scripts_ tab, selecting the sync script, and checking in the sync script settings if it supports webhooks and, if so, which webhook subscription(s). 
+You can check if a sync supports external webhooks by selecting your integration, going to the _Scripts_ tab, selecting the sync script, and checking in the sync script settings if it supports webhooks and, if so, which webhook subscription(s).
 
 <Tip>
 **Questions, problems, feedback?** Please reach out in the [Slack community](https://nango.dev/slack).
