+++ docs-v2/integrate/guides/sync-data-from-an-api.mdx
@@ -17,16 +17,16 @@ Select your integration in the _Integrations_ tab, and navigate to the _Scripts_
 Nango will automatically sync the corresponding [records](/understand/concepts/syncs) in the background for each relevant connection.
 
 <Tip>
-Is there no template for your API? Or none matching your exact use case? 
+Is there no template for your API? Or none matching your exact use case?
 
 Learn more about how to [build a custom integration](/customize/guides/create-a-custom-integration), [extend a template](/customize/guides/extend-an-integration-template) or [request custom integrations from Nango experts](/host/managed-integrations).
 </Tip>
 
 # Listen for webhooks from Nango
 
-Nango sends [webhook](/understand/concepts/webhooks) notifications to your backend whenever new data is available for a connection & sync combination. 
+Nango sends [webhook](/understand/concepts/webhooks) notifications to your backend whenever new data is available for a connection & sync combination.
 
-To set this up, go to the _Environment Settings_ tab and configure a _Webhook URL_ to which Nango will send notifications. 
+To set this up, go to the _Environment Settings_ tab and configure a _Webhook URL_ to which Nango will send notifications.
 
 The webhook from Nango is a POST request with the following body:
 ```json
@@ -35,7 +35,7 @@ The webhook from Nango is a POST request with the following body:
     "providerConfigKey": "<INTEGRATION-ID>",
     "syncName": "<SYNC-NAME>",
     "model": "<MODEL-NAME>",
-    "responseResults": { "<MODEL-NAME>": { "added": 123, "updated": 123, "deleted": 123 } },
+    "responseResults": { "added": 123, "updated": 123, "deleted": 123 },
     "syncType": "INITIAL" | "INCREMENTAL",
     "modifiedAfter": "<TIMESTAMP>"
 }
@@ -66,7 +66,7 @@ curl -G https://api.nango.dev/records \
   --header 'Connection-Id: <connectionId-in-webhook-payload>' \
   --data-urlencode 'model=<model-in-webhook-payload>' \
   --data-urlencode 'modified_after=<modifiedAfter-in-webhook-payload>' \
-  
+
 ```
 
 </Tab>
@@ -89,7 +89,7 @@ const result = await nango.listRecords({
 
 </Tabs>
 
-This returns an array of records conforming to the specified data model. 
+This returns an array of records conforming to the specified data model.
 
 Each record contains useful metadata automatically generated by Nango:
 
@@ -116,7 +116,7 @@ Each record contains useful metadata automatically generated by Nango:
 
 ### Cursor-based synchronization
 
-In practice, webhook notifications can be missed, and relying solely on the webhook payload to fetch modified records can cause you to miss some updates. 
+In practice, webhook notifications can be missed, and relying solely on the webhook payload to fetch modified records can cause you to miss some updates.
 
 A more reliable way of keeping track of how far you've synced records (for each connection & sync combination) is to rely on record **cursors**.
 
@@ -137,7 +137,7 @@ curl -G https://api.nango.dev/records \
   --header 'Connection-Id: <connectionId-in-webhook-payload>' \
   --data-urlencode 'model=<model-in-webhook-payload>' \
   --data-urlencode 'cusor=<cursor-of-last-fetched-record>' \
-  
+
 ```
 
 </Tab>
@@ -160,7 +160,7 @@ const result = await nango.listRecords({
 
 </Tabs>
 
-So, the overall logic for cursor-based synchronization should be: 
+So, the overall logic for cursor-based synchronization should be:
 1. Receive a webhook notification from Nango
 2. Query your database for the cursor of the last-fetched record
 3. Fetch the modified records (passing the cursor)
@@ -170,11 +170,11 @@ So, the overall logic for cursor-based synchronization should be:
 
 # Write back to APIs (2-way syncing)
 
-Write back to APIs with actions ([step-by-step guide](/integrate/guides/perform-workflows-with-an-api)) or proxy requests ([step-by-step guide](/integrate/guides/proxy-requests-to-an-api)). 
+Write back to APIs with actions ([step-by-step guide](/integrate/guides/perform-workflows-with-an-api)) or proxy requests ([step-by-step guide](/integrate/guides/proxy-requests-to-an-api)).
 
 # Troubleshoot errors & monitor
 
-Navigate to the _Activity_ tab to inspect potential errors & monitor sync executions. 
+Navigate to the _Activity_ tab to inspect potential errors & monitor sync executions.
 
 <Tip>
 **Questions, problems, feedback?** Please reach out in the [Slack community](https://nango.dev/slack).
